Track B: f32 draw lanes, f64 aggregation, across all three backends
Random-variable lanes become f32 (interpreter register file, Cranelift JIT
SSA, emitted wasm kernel); deterministic values and every reducer/estimate
stay f64. All three backends remain bit-identical to each other (full-batch
bitwise conformance).

Draws are now pair-shared: an f32 uniform needs 24 bits and one squares64
yields 48 consumable ones, so unif/normal/exp/geometric hash once per lane
PAIR (even lane takes the low 24 bits, odd the high 24). unif_int keeps its
per-lane 48-bit Lemire draw -- 24-bit Lemire would put its bias at
count/2^24. The certified stream is unchanged by this: each source still
walks counters 0,1,2,... spending all 48 bits in order, so C0's 1 TB
criterion-8 evidence carries over (argued in tools/rng-cert/RESULTS.md).

f32 Box-Muller uses r = sqrt(-2*ln(1 - u1)) instead of the half-ulp nudge:
on the 2^-24 grid 1-u1 is exactly representable in [2^-24, 1], making the
ln(0) guard structural. approx.rs gains shared f32 ln/sin/cos polynomials
that both emitters inline. Rotation's Gram-Schmidt stays in f64 scratch so
orthonormality is bound by f32 storage, not f32 arithmetic -- turboquant
still reproduces the paper's distortion table.

Corpus gate: 3938 ms vs 4130 post-Squares (-4.6%), -9.5% vs the pre-PREGPU
baseline. fill_uniform 540 -> 1039 M/s (1.92x); fill_normal is flat (90 vs
86) because Box-Muller is transcendental-latency bound, not hash bound.

New language boundaries, documented in LANG.md and enforced/pinned by tests:
unif_int bounds capped at +/-2^24; a distribution's spread must not be below
~1e-7 of its location (normal(1e8, 1) is no longer representable);
exp/geometric tails truncate at ~16.6/rate; 2^32 draws per forcing.

Also correct a stale guarantee: counter keying made sample_n_par draw the
IDENTICAL stream as sample_n (a chunk is just a lane range), which the docs
still described as "same distribution, different draws". Now tested.

+++ crates/noise-core/src/backend.rs
@@ -195,7 +195,13 @@ pub trait Runner {
     /// Produce the next batch of draws and return the **first `len`** root samples (`len <=
     /// batch_cap()`). Advances the lane cursor by a *fixed* `batch_cap()` per call (independent of
     /// `len`), so the draw stream doesn't depend on how a final partial batch is sliced.
-    fn next_batch(&mut self, len: usize) -> &[f64];
+    ///
+    /// Lanes are **f32** (PLAN-PREGPU Track B). This one signature is the whole type change: a
+    /// draw is a Monte-Carlo sample whose `O(1/√N)` standard error dwarfs f32's ~1e-7 rounding, so
+    /// the transport type can be narrow — while everything that becomes a *reported estimate*
+    /// (`reduce`'s Σx/Σx² accumulators, quantile interpolation, the public draw vectors) widens to
+    /// f64 at this boundary and stays there. Narrow lanes, wide sums.
+    fn next_batch(&mut self, len: usize) -> &[f32];
 
     /// Maximum `len` accepted by [`next_batch`](Runner::next_batch) — the backend's column width.
     fn batch_cap(&self) -> usize;
@@ -224,7 +230,7 @@ impl Program for InterpProgram {
         // Allocate this worker's column file once; reused across its batches. Array registers
         // (`Inst::Permutation`) get their own `n × BATCH` buffers, sized by `Program::arrays`.
         let regs = (0..self.inner.n_regs)
-            .map(|_| vec![0.0f64; BATCH].into_boxed_slice())
+            .map(|_| vec![0.0f32; BATCH].into_boxed_slice())
             .collect();
         let arrs = alloc_arrays(&self.inner);
         // Placeholder key/lane; the driver calls `position` before the first batch.
@@ -240,10 +246,10 @@ impl Program for InterpProgram {
 
 /// Allocate a worker's array-register file (one lane-major `n × BATCH` buffer per array register;
 /// see `bytecode::Program::arrays`). Empty for the common no-array program.
-fn alloc_arrays(prog: &ByteProgram) -> Vec<Box<[f64]>> {
+fn alloc_arrays(prog: &ByteProgram) -> Vec<Box<[f32]>> {
     prog.arrays
         .iter()
-        .map(|&n| vec![0.0f64; n as usize * BATCH].into_boxed_slice())
+        .map(|&n| vec![0.0f32; n as usize * BATCH].into_boxed_slice())
         .collect()
 }
 
@@ -251,8 +257,8 @@ fn alloc_arrays(prog: &ByteProgram) -> Vec<Box<[f64]>> {
 /// draw key + lane cursor.
 struct InterpRunner {
     prog: Arc<ByteProgram>,
-    regs: Vec<Box<[f64]>>,
-    arrs: Vec<Box<[f64]>>,
+    regs: Vec<Box<[f32]>>,
+    arrs: Vec<Box<[f32]>>,
     key: Key,
     lane: u32,
 }
@@ -263,7 +269,7 @@ impl Runner for InterpRunner {
         self.lane = lane;
     }
 
-    fn next_batch(&mut self, len: usize) -> &[f64] {
+    fn next_batch(&mut self, len: usize) -> &[f32] {
         // Fill the full BATCH (lane consumption is constant per call), then slice to `len`.
         run_batch(&self.prog, &mut self.regs, &mut self.arrs, self.key, self.lane);
         self.lane = self.lane.wrapping_add(BATCH as u32);
@@ -294,8 +300,9 @@ pub trait JointRunner {
     fn next_batch(&mut self);
 
     /// Root `j`'s column of the current batch (`batch_cap()` lanes; the driver slices a final
-    /// partial batch itself). Lane `i` across all columns is one *joint* draw.
-    fn col(&self, j: usize) -> &[f64];
+    /// partial batch itself). Lane `i` across all columns is one *joint* draw. f32, like
+    /// [`Runner::next_batch`].
+    fn col(&self, j: usize) -> &[f32];
 
     /// Lanes per batch — the backend's column width.
     fn batch_cap(&self) -> usize;
@@ -310,7 +317,7 @@ struct InterpJointProgram {
 impl JointProgram for InterpJointProgram {
     fn runner(&self) -> Box<dyn JointRunner> {
         let buf = (0..self.inner.n_regs)
-            .map(|_| vec![0.0f64; BATCH].into_boxed_slice())
+            .map(|_| vec![0.0f32; BATCH].into_boxed_slice())
             .collect();
         let arrs = alloc_arrays(&self.inner);
         // Placeholder key/lane; the driver calls `position` before the first batch.
@@ -329,8 +336,8 @@ impl JointProgram for InterpJointProgram {
 struct InterpJointRunner {
     prog: Arc<ByteProgram>,
     regs: Vec<usize>,
-    buf: Vec<Box<[f64]>>,
-    arrs: Vec<Box<[f64]>>,
+    buf: Vec<Box<[f32]>>,
+    arrs: Vec<Box<[f32]>>,
     key: Key,
     lane: u32,
 }
@@ -346,7 +353,7 @@ impl JointRunner for InterpJointRunner {
         self.lane = self.lane.wrapping_add(BATCH as u32);
     }
 
-    fn col(&self, j: usize) -> &[f64] {
+    fn col(&self, j: usize) -> &[f32] {
         &self.buf[self.regs[j]]
     }
